feat : Long 타입 변경 / StoreProduct Tag와 Category enum으로 변경

+++ backEnd/app/src/main/java/com/kh/app/store/entity/StoreProductImageEntity.java
@@ -23,7 +23,7 @@ public class StoreProductImageEntity extends BaseEntity {
     private StoreProductEntity product;
 
     @Column(name = "SORT_ORDER")
-    private Integer sortOrder;
+    private Long sortOrder;
 
     @Column(name = "IMAGE_REPRESENT_YN", nullable = false, length = 1)
     @Builder.Default
@@ -39,7 +39,7 @@ public class StoreProductImageEntity extends BaseEntity {
             StoreProductEntity product,
             MultipartFile file,
             String changedName,
-            Integer sortOrder,
+            Long sortOrder,
             String imageRepresentYn
     ) {
         return StoreProductImageEntity.builder()
